libwasi: rename a few things
