+++ libwasi/CMakeLists.txt
@@ -3,7 +3,7 @@
 set(lib_wasi_sources
 	"wasi.c"
 	"wasi_fdtable.c"
-	"wasi_host_dirop.c"
+	"wasi_host_pathop.c"
 	"wasi_host_fdop.c"
 	"wasi_host_fdop_readdir.c"
 	"wasi_host_subr.c"
