+++ libwasi/wasi.c
@@ -71,8 +71,8 @@
 #include "util.h"
 #include "vec.h"
 #include "wasi.h"
-#include "wasi_dirop.h"
-#include "wasi_fdop.h"
+#include "wasi_host_pathop.h"
+#include "wasi_host_fdop.h"
 #include "wasi_host_subr.h"
 #include "wasi_impl.h"
 #include "wasi_utimes.h"
