+++ libwasi/wasi_host_pathop.h
